Updated to account for Windows guests.
When building new Windows Packer templates I have run into an issue with
WinRM communication which has been resolved with the WinRM settings. I
also ran into an issue where the default virtio network driver would not
work with VirtualBox guest additions. So, to get around this, we are
pinning the nic_type to 82540EM which works.